Rename store2.rs to store.rs
